+++ R/utils.R
@@ -154,6 +154,58 @@ random_name <- function(prefix = "") {
   paste0(prefix, "odbc_", name)
 }
 
+# error handling ---------------------------------------------------------------
+# `call` looks to the "user-called" function rather than the usual `caller_env(n)`.
+# this function is called from C++, and parent.frame() (and thus caller_env())
+# is unable to traverse across the Rcpp code to find the caller env; setting
+# `call = caller_env(n)` with n of 0/1/2/3 gives no error context (#789).
+rethrow_database_error <- function(msg, call = trace_back()$call[[1]]) {
+  tryCatch(
+    res <- parse_database_error(msg),
+    error = function(e) cli::cli_abort(msg, call = call)
+  )
+
+  cli::cli_abort(
+    c(
+      "!" = "ODBC failed with error {res$cnd_context_code} from \\
+             {.field {paste0(res$cnd_context_driver, collapse = '')}}.",
+      set_names(res$cnd_body, nm = c("x", rep("*", length(res$cnd_body) - 1))),
+      "i" = "From {.file {res$cnd_context_nanodbc}}."
+    ),
+    call = call
+  )
+}
+
+parse_database_error <- function(msg) {
+  # Split nanodbc's context, error code, and message returned from the database
+  cnd_msg <- strsplit(msg, "\n", fixed = TRUE)[[1]]
+
+  # Parse out nanodbc/nanodbc.cpp:<line#> and 5-char error code
+  cnd_context <- strsplit(cnd_msg[1], ": ")[[1]]
+  cnd_context_nanodbc <- cnd_context[1]
+  cnd_context_code <- cnd_context[2]
+
+  # In error returned from the database, find the square-bracketed context
+  # on driver and driver manager
+  cnd_context_driver <- gregexpr("(\\[.*?\\])", cnd_msg[-1])
+  cnd_context_driver <- regmatches(cnd_msg[-1], cnd_context_driver)[[1]]
+
+  # Remove the square-bracketed context from the database error
+  cnd_body <- Reduce(
+    function(p, x) gsub(p, "", x, fixed = TRUE),
+    cnd_context_driver,
+    cnd_msg[-1],
+    right = TRUE
+  )
+
+  list(
+    cnd_context_nanodbc = cnd_context_nanodbc,
+    cnd_context_code = cnd_context_code,
+    cnd_context_driver = cnd_context_driver,
+    cnd_body = cnd_body
+  )
+}
+
 # check helpers for common odbc arguments --------------------------------------
 check_row.names <- function(row.names, call = caller_env()) {
   if (is.null(row.names) || is_scalar_logical(row.names) || is_string(row.names)) {